Created some new helpers, moved old ones to core
