+++ src/FunderMaps.Core/Extensions/StringExtensions.cs
@@ -14,8 +14,15 @@ namespace FunderMaps.Webservice.Utility
         /// <param name="s">The <see cref="string"/> to check</param>
         public static void ThrowIfNullOrEmpty(this string s)
         {
-            if (s == null) { throw new ArgumentNullException(nameof(s)); }
-            if (string.IsNullOrEmpty(s)) { throw new ArgumentNullException(nameof(s)); }
+            if (s == null)
+            {
+                throw new ArgumentNullException(nameof(s));
+            }
+
+            if (string.IsNullOrEmpty(s))
+            {
+                throw new ArgumentNullException(nameof(s));
+            }
         }
     }
 }
